asYouType: properly defined country phone code on countriness reset
Country metadata is an array and not an object, thus while running
reset_countriness, the ountry_phone_code was alway set to undefined,
instead of the default one.

This issue seems to have no impact on the lib behavior, however, this
is annoying for custom behavior as always returning international
partial number.