Fix argument type of `isSupportedCountry` helper
Since the whole point of the helper is to check whether an arbitrary string is a valid country code supported by the library, we should accept any string here. For instance, the example code `isSupportedCountry("XX")` causes a compiler error, becaue "XX" is not part of the `CountryCode` union type.